+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 Changelog
 =========
+
+## Unreleased
+### Added
+- BaseAccessor service query trait
+
 ## 1.0.0-rc.12 - 2018-02-22
 ### Added
 - NotFoundException to indicate a generic not found error